refactor

+++ util.c
@@ -1,5 +1,5 @@
 #include "util.h"
 
-int collision(int x1, int y1, int h1, int w1, int x2, int y2, int w2, int h2){
+double collision(double x1, double y1, double h1, double w1, double x2, double y2, double w2, double h2){
     return (MAX(x1, x2) < MIN(x1 + w1, x2 + w2)) && (MAX(y1, y2) < MIN(y1 + h1, y2 + h2));
 }
